[StoryDetail] Extract widgets to separate classes #64
